enable to get several endpoints data
